+++ lib/jazzy/doc_builder.rb
@@ -30,8 +30,7 @@ module Jazzy
     def self.doc_structure_for_docs(docs)
       docs.map do |doc|
         children = doc.children
-                   .sort_by(&:name)
-                   .sort_by(&:nav_order)
+                   .sort_by { |c| [c.nav_order, c.name] }
                    .map do |child|
           { name: child.name, url: child.url }
         end
@@ -188,7 +187,8 @@ module Jazzy
     end
 
     def self.copy_assets(destination)
-      FileUtils.cp_r(Config.instance.assets_directory.children, destination)
+      assets_directory = Config.instance.theme_directory + 'assets'
+      FileUtils.cp_r(assets_directory.children, destination)
       Pathname.glob(destination + 'css/**/*.scss').each do |scss|
         contents = scss.read
         css = Sass::Engine.new(contents, syntax: :scss).render
@@ -207,6 +207,7 @@ module Jazzy
       doc = Doc.new # Mustache model instance
       doc[:name] = source_module.name
       doc[:overview] = ReadmeGenerator.generate(source_module)
+      doc[:custom_head] = Config.instance.custom_head
       doc[:doc_coverage] = source_module.doc_coverage unless
         Config.instance.hide_documentation_coverage
       doc[:structure] = source_module.doc_structure
@@ -312,6 +313,7 @@ module Jazzy
       end
 
       doc = Doc.new # Mustache model instance
+      doc[:custom_head] = Config.instance.custom_head
       doc[:doc_coverage] = source_module.doc_coverage unless
         Config.instance.hide_documentation_coverage
       doc[:name] = doc_model.name