+++ lib/jazzy/jazzy_markdown.rb
@@ -44,7 +44,6 @@ module Jazzy
                             Version
                             Warning).freeze
 
-    # rubocop:disable RegexpLiteral
     SPECIAL_LIST_TYPE_REGEX = %r{
       \A\s* # optional leading spaces
       (<p>\s*)? # optional opening p tag
@@ -52,7 +51,6 @@ module Jazzy
       (#{SPECIAL_LIST_TYPES.map(&Regexp.method(:escape)).join('|')})
       [\s:] # followed by either a space or a colon
     }ix
-    # rubocop:enable RegexpLiteral
 
     ELIDED_LI_TOKEN = '7wNVzLB0OYPL2eGlPKu8q4vITltqh0Y6DPZf659TPMAeYh49o'.freeze
 
@@ -68,10 +66,12 @@ module Jazzy
     end
 
     def render_aside(type, text)
-      %(<div class="aside aside-#{type.underscore.tr('_', '-')}">
-          <p class="aside-title">#{type.underscore.humanize}</p>
-          #{text}
-        </div>)
+      <<-HTML
+<div class="aside aside-#{type.underscore.tr('_', '-')}">
+    <p class="aside-title">#{type.underscore.humanize}</p>
+    #{text}
+</div>
+      HTML
     end
 
     def list(text, list_type)