+++ lib/jazzy/sourcekitten.rb
@@ -223,24 +223,13 @@ module Jazzy
         doc['key.parsed_declaration'] || doc['key.doc.declaration'],
         Config.instance.objc_mode ? 'objc' : 'swift',
       )
-      declaration.abstract = comment_from_doc(doc)
+      declaration.abstract = Jazzy.markdown.render(doc['key.doc.comment'] || '')
       declaration.discussion = ''
       declaration.return = make_paragraphs(doc, 'key.doc.result_discussion')
 
       declaration.parameters = parameters(doc)
 
       @documented_count += 1
-    end
-
-    def self.comment_from_doc(doc)
-      swift_version = Config.instance.swift_version.to_f
-      comment = doc['key.doc.comment'] || ''
-      if swift_version < 2
-        # comment until first ReST definition
-        matches = /^\s*:[^\s]+:/.match(comment)
-        comment = comment[0...matches.begin(0)] if matches
-      end
-      Jazzy.markdown.render(comment)
     end
 
     def self.make_substructure(doc, declaration)
@@ -330,15 +319,19 @@ module Jazzy
     # Two declarations get merged if they have the same deduplication key.
     def self.deduplication_key(decl)
       if decl.type.swift_extensible? || decl.type.swift_extension?
-        [decl.usr]
+        [decl.usr, decl.name]
+      elsif decl.type.objc_class? || decl.type.objc_category?
+        name, _ = decl.objc_category_name || decl.name
+        [name, :objc_class_and_categories]
       else
-        [decl.usr, decl.type.kind]
+        [decl.usr, decl.name, decl.type.kind]
       end
     end
 
+    # rubocop:disable Metrics/MethodLength
     # Merges all of the given types and extensions into a single document.
     def self.merge_declarations(decls)
-      extensions, typedecls = decls.partition { |d| d.type.swift_extension? }
+      extensions, typedecls = decls.partition { |d| d.type.extension? }
 
       if typedecls.size > 1
         warn 'Found conflicting type declarations with the same name, which ' \
@@ -351,6 +344,13 @@ module Jazzy
         merge_default_implementations_into_protocol(typedecl, extensions)
         mark_members_from_protocol_extension(extensions)
         extensions.reject! { |ext| ext.children.empty? }
+      elsif typedecl && typedecl.type.objc_class?
+        # Mark children merged from categories with the name of category
+        # (unless they already have a mark)
+        extensions.each do |ext|
+          _, category_name = ext.objc_category_name
+          ext.children.each { |c| c.mark.name ||= category_name }
+        end
       end
 
       decls = typedecls + extensions
@@ -362,6 +362,7 @@ module Jazzy
         end
       end
     end
+    # rubocop:enable Metrics/MethodLength
 
     # If any of the extensions provide default implementations for methods in
     # the given protocol, merge those members into the protocol doc instead of